+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 v0.8.5
 - [BUGFIX] fixed issue where budget percentages are displayed wrong
 - [BUGFIX] in show-all order date by date value not by string value
+- [BUGFIX] ask if categories should be created or not
+- [UI] minor changes
 -