+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # CHANGELOG
 
-v0.9.0
+v0.9.1
 - [TASK] added buttons for cycling throw months in the budget overview view
+- [UPDATE] Bump maven-compiler-plugin from 3.9.0 to 3.10.0
+- [TASK] init csv export